% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_postprodn_fns.R
\name{as.mcmc.list.angmcmc}
\alias{as.mcmc.list.angmcmc}
\title{Create an mcmc.list object from an angmcmc object}
\usage{
\method{as.mcmc.list}{angmcmc}(x, ...)
}
\arguments{
\item{x}{angmcmc object}

\item{...}{unused}
}
\description{
Create an mcmc.list object from an angmcmc object
}
\examples{
# first fit a vmsin mixture model
# illustration only - more iterations needed for convergence
fit.vmsin.20 <- fit_vmsinmix(tim8, ncomp = 3, n.iter =  20,
                             n.chains = 1)

# now convert it to mcmc.list
library(coda)
fit.vmsin.20.mcmc <- as.mcmc.list(fit.vmsin.20)
}
