% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_class.R
\name{bambi}
\alias{bambi}
\title{bambi}
\description{
motivation explanation
queries WGS and SS 10x BAMs
}
\section{Arguments}{

\describe{
  \item{p}{A \code{process} object.}
  \item{command}{Character scalar, the command to run. It will be
    escaped via \code{\link[base]{shQuote}}.}
  \item{args}{Character vector, arguments to the command. The will be
    escaped via \code{\link[base]{shQuote}}.}
  \item{commandline}{A character scalar, a full command line.
    No escaping will be performed on it.}
  \item{stdout}{What to do with the standard output. Possible values:
    \code{FALSE}: discard it; a string, redirect it to this file,
    \code{TRUE}: redirect it to a temporary file.}
  \item{stdout}{What to do with the standard error. Possible values:
    \code{FALSE}: discard it; a string, redirect it to this file,
    \code{TRUE}: redirect it to a temporary file.}
  \item{grace}{Grace pediod between the TERM and KILL signals, in
    seconds.}
  \item{...}{Extra arguments are passed to the
    \code{\link[base]{readLines}} function.}
}
}

\section{Details}{

\code{$grab_bx()} grab BX tags (if exist)

grab_bx = function(query, data.table = FALSE, verbose = FALSE, mc.cores = 1)

\code{$grab_cb()} grab CB tags (if exist)

grab_cb = function(query, data.table = FALSE, verbose = FALSE, mc.cores = 1)

\code{$grab_ub()} grab CB tags (if exist)

grab_ub = function(query, data.table = FALSE, verbose = FALSE, mc.cores = 1)

\code{$fetch_by_tag()} grab CB tags (if exist)

fetch_by_tag = function(tag, query, data.table = FALSE, verbose = FALSE, mc.cores = 1)
}

