% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_index}
\alias{check_index}
\title{checks the LMDB indices to check whether these exist}
\usage{
check_index(lmdb_path, index)
}
\arguments{
\item{lmdb_path}{List output from query_bam_index()}

\item{index}{character string LMDB index to check if exists}
}
\value{
boolean TRUE if LMDB has index
}
\description{
Checks the LMDB indices to check whether these exist
}
