% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_vmsin_fns.R
\name{fit_vmsinmix}
\alias{fit_vmsinmix}
\title{Fitting bivariate von Mises sine model mixtures using MCMC}
\usage{
fit_vmsinmix(...)
}
\arguments{
\item{...}{arguments (other than \code{model}) passed to \link{fit_angmix}}
}
\description{
Fitting bivariate von Mises sine model mixtures using MCMC
}
\details{
Wrapper for \link{fit_angmix} with \code{model = "vmsin"}
}
\examples{
# illustration only - more iterations needed for convergence
fit.vmsin.20 <- fit_vmsinmix(tim8, ncomp = 3, n.iter =  20,
                             n.chains = 1)
fit.vmsin.20

}
