% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_wnorm2_fns.R
\name{fit_wnorm2mix}
\alias{fit_wnorm2mix}
\title{Fitting bivariate wrapped normal model mixtures using MCMC}
\usage{
fit_wnorm2mix(...)
}
\arguments{
\item{...}{arguments (other than \code{model}) passed to \link{fit_angmix}}
}
\description{
Fitting bivariate wrapped normal model mixtures using MCMC
}
\details{
Wrapper for \link{fit_angmix} with \code{model = "wnorm2"}.
}
\examples{
# illustration only - more iterations needed for convergence
fit.wnorm2.10 <- fit_wnorm2mix(tim8, ncomp = 3, n.iter =  10,
                             n.chains = 1)
fit.wnorm2.10

}
