% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_wnorm_fns.R
\name{fit_wnormmix}
\alias{fit_wnormmix}
\title{Fitting univariate wrapped normal mixtures using MCMC}
\usage{
fit_wnormmix(...)
}
\arguments{
\item{...}{arguments (other than \code{model}) passed to \link{fit_angmix}}
}
\description{
Fitting univariate wrapped normal mixtures using MCMC
}
\details{
Wrapper for \link{fit_angmix} with \code{model = "wnorm"}.
}
\examples{
# illustration only - more iterations needed for convergence
fit.wnorm.20 <- fit_wnormmix(wind$angle, ncomp = 3, n.iter =  20,
                             n.chains = 1)
fit.wnorm.20

}
