% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countCigar}
\alias{countCigar}
\alias{bamflag}
\title{Count bases in cigar string}
\usage{
countCigar(cigar)

bamflag(reads)
}
\arguments{
\item{cigar}{character vector of cigar strings}

\item{reads}{data.table holding the reads}
}
\value{
a 4-column, length(cigar)-row matrix with the total counts for each type

matrix of bits from BAM flags
}
\description{
Counts the total number of bases, per cigar, that fall into D, I, M, S categories.
countCigar() makes no distinction between, for instance 1S2M2S, 2S2M1S, or 3S2M

Shortcut function in create_granges
}
