% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_outputs}
\alias{parse_outputs}
\title{parses output of BAM queries into GRanges with parsed CIGAR strings}
\usage{
parse_outputs(out)
}
\arguments{
\item{out}{List output from query_bam_index()}
}
\value{
GRanges with parsed CIGAR
}
\description{
Internal function to parse CIGAR and output GRanges
}
