% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_postprodn_fns.R
\name{select_chains}
\alias{select_chains}
\title{Select chains from angmcmc objects}
\usage{
select_chains(object, chain.no, ...)
}
\arguments{
\item{object}{angular MCMC object.}

\item{chain.no}{labels of chains to be retained in the final sample. If missing,
all chains are used.}

\item{...}{unused}
}
\value{
Returns another angmcmc object with only selected chains passed through \code{chain.no}
}
\description{
Select chains from angmcmc objects
}
\examples{
# illustration only - more iterations needed for convergence
fit.vmsin.20 <- fit_vmsinmix(tim8, ncomp = 3, n.iter =  20,
                             L = c(10, 12), chains_parallel = FALSE,
                             n.chains = 2)
fit.vmsin.20
fit.vmsin.20.1 <- select_chains(fit.vmsin.20, 1)
fit.vmsin.20.1

}
