% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tim8}
\alias{tim8}
\title{Backbone Dihedral Angles of Triose Phosphate Isomerase (8TIM)}
\format{A data frame with 490 rows and 2 variables (backbone dihedral angles) phi and psi.}
\source{
8TIM PDB file: \url{http://www.rcsb.org/pdb/explore.do?structureId=8tim}.

DSSP software: \url{http://swift.cmbi.ru.nl/gv/dssp/}.
}
\usage{
data(tim8)
}
\description{
A dataset consisting of 490 pairs of backbone dihedral angles (in radian scale \eqn{[0, 2\pi)} )
\eqn{(\phi, \psi)} for the protein Triose Phosphate Isomerase (8TIM). The angles were obtained first by using
the DSSP software on the PDB file for 8TIM to get the backbone angles (in degrees),
and then by converting all angles into radians.
}
\keyword{datasets}
