% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMBI-package.R
\name{BAMBI}
\alias{BAMBI}
\title{\code{BAMBI}: An R package for Bivariate Angular Mixture Models}
\description{
\code{BAMBI} is an R package that provides functions for fitting
(using Bayesian methods) and simulating mixtures of univariate
and bivariate angular distributions. See the reference for a
detailed description of the functionalities of \code{BAMBI}.
}
\references{
Chakraborty, S., & Wong, S. W. (2021). BAMBI: An R package for
fitting bivariate angular mixture models. \emph{Journal of Statistical Software},
99 (11), 1-69. \doi{10.18637/jss.v099.i11}
}
