% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_vm_fns.R
\name{fit_vmmix}
\alias{fit_vmmix}
\title{Fitting univariate von Mises mixtures using MCMC}
\usage{
fit_vmmix(...)
}
\arguments{
\item{...}{arguments (other than \code{model}) passed to \link{fit_angmix}}
}
\description{
Fitting univariate von Mises mixtures using MCMC
}
\details{
Wrapper for \link{fit_angmix} with \code{model = "vm"}.
}
\examples{
# illustration only - more iterations needed for convergence
fit.vm.20 <- fit_vmmix(wind$angle, ncomp = 3, n.iter =  20,
                       n.chains = 1)
fit.vm.20
}
